Extract harness event logging
